refactor: adds multiple launch configurations
- The `Local Frege LSP` launch config allows to run a local version.
 Just set the `LOCAL_FREGE_SERVER_PATH` env variable.
- The `Local Frege LSP with Java Debugging` launch config additionally
allows to debug the local lsp server end-to-end inside vscode. Just
set a breakpoint in the Java code, start the local client and server,
attach the java debugger on port 6008 (currently hardcoded) and trigger
a LSP feature in the client. The breakpoint should be automatically hit
(like magic, I know).